Update error logging and migration column type

+++ lib/node-storage/migrations/postgres/20260129120000_add_blockfrost_page.sql
@@ -1,2 +1,2 @@
 -- Add blockfrost_page column to dlt_cursor table for pagination tracking
-ALTER TABLE dlt_cursor ADD COLUMN blockfrost_page INT;
+ALTER TABLE dlt_cursor ADD COLUMN blockfrost_page INT8;
